# public/swagger/openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel",
        "version": "1.0.0"
    },
    "paths": {
        "/api/category/category?id={id}": {
            "get": {
                "summary": "Gets a category based on id",
                "operationId": "App\\Http\\Controllers\\CategoryController::get",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Parameter with mutliple examples",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "examples": {
                            "int": {
                                "summary": "An int value.",
                                "value": "1"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
        "/api/category/category": {
            "put": {
                "summary": "Update a category",
                "operationId": "App\\Http\\Controllers\\CategoryController::update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "Parameter with mutliple examples",
                        "required": true
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Parameter with mutliple examples",
                        "required": true
                    },
                    {
                        "name": "parent",
                        "in": "query",
                        "description": "Parameter with mutliple examples",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "post": {
                "summary": "Create a category",
                "operationId": "App\\Http\\Controllers\\CategoryController::create",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Parameter with mutliple examples",
                        "required": true
                    },
                    {
                        "name": "parent",
                        "in": "query",
                        "description": "Parameter with mutliple examples",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "delete": {
                "summary": "Delete a category",
                "operationId": "App\\Http\\Controllers\\CategoryController::delete",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "Parameter with mutliple examples",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/comment/comment?id={id}": {
            "get": {
                "summary": "Gets a poscommentt based on id",
                "operationId": "App\\Http\\Controllers\\CommentController::get",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Parameter with mutliple examples",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "examples": {
                            "int": {
                                "summary": "An int value.",
                                "value": "1"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
    }
}